progress

+++ src/player/components.rs
@@ -5,7 +5,7 @@ pub struct Player {
    pub speed: f32,
    pub last_position: Vec3,
    pub damage_cooldown: Timer,
-   pub experience: f32,
+   pub experience: i32,
    pub level: u32,
    pub kills: u32,
 }
